Refactor installation tasks into Rake

+++ lib/tasks/install.rake
@@ -1,25 +1,38 @@
 namespace :install do
-  desc "Install certain config files and binaries to bootstrap DOTS"
-  task :configuration => ['install:binaries'] do
-    %w(zshrc zshenv vimrc).each do |filename|
-      sh "ln -s ~/.dots/config/#{filename} ~/.#{filename}"
-    end
+  task :links do
+    sh 'ln -s ~/.dots/bin ~/bin' unless Dir.exists? "~/bin"
+    sh "ln -s ~/.dots/etc ~/etc" unless Dir.exists? "~/etc"
   end
 
-  task :binaries do
-    sh 'ln -s ~/.dots/bin ~/bin'
+  task :configuration do
+    sh 'dots update'
+    sh 'dots reload'
   end
 
-  desc "Install vendored code in the DOTS folder"
   task :submodules do
     sh 'cd ~/.dots && git submodule init && git submodule update'
+  end
+
+  task :applications do
+    application_config_path = File.expand_path "~/.dots/applications"
+    applications = File.read(application_config_path).split("\n").join(' ')
+    install_command = if `cat /proc/version` =~ /Linux/
+      "apt-get install -y #{names}"
+    else
+      sh %{ ruby -e "$(curl -fsSL https://raw.github.com/mxcl/homebrew/go)" } \
+        unless `which brew` =~ /brew/
+
+      "brew install #{names}"
+    end
+
+    sh install_command
   end
 end
 
 desc "Install DOTS to this user's home directory"
 task :install => %w(
+  install:links
   install:configuration
-  install:binaries
   install:submodules
   install:applications
 )
